requirements.txt
# Trace Agricultural AI System — requirements.txt
# Core runtime and AI dependencies

# ── Core scientific stack ──────────────────────────────────────────────────
numpy>=1.26,<2.0
scipy>=1.12,<2.0
pandas>=2.2,<3.0
Pillow>=10.3,<11.0

# ── Geospatial / Satellite ─────────────────────────────────────────────────
rasterio>=1.3,<2.0
shapely>=2.0,<3.0
pyproj>=3.6,<4.0
fiona>=1.9,<2.0
geopandas>=0.14,<1.0
sentinelsat>=1.3,<2.0
earthengine-api>=0.1.390,<1.0

# ── ML / forecasting / AI ──────────────────────────────────────────────────
prophet>=1.1,<2.0
torch>=2.2,<3.0
scikit-learn>=1.4,<2.0
faiss-cpu>=1.8,<2.0
sentence-transformers>=2.7,<3.0
google-generativeai>=0.7,<1.0
langchain>=0.2,<1.0
langgraph>=0.1,<1.0

# ── Database (Historical & Farmer DB) ──────────────────────────────────────
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
aiosqlite>=0.20,<1.0
asyncpg>=0.29,<1.0
alembic>=1.13,<2.0
python-dateutil>=2.9,<3.0

# ── Backend ────────────────────────────────────────────────────────────────
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.7,<3.0
python-multipart>=0.0.9,<1.0
redis>=5.0,<6.0
apscheduler>=3.10,<4.0

# ── Communications ─────────────────────────────────────────────────────────
twilio>=9.1,<10.0

# ── Config / Utility ───────────────────────────────────────────────────────
pyyaml>=6.0,<7.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
tenacity>=8.3,<10.0
structlog>=24.1,<25.0

# ── Testing ────────────────────────────────────────────────────────────────
pytest>=8.1,<9.0
pytest-asyncio>=0.23,<1.0
pytest-cov>=5.0,<7.0
